Fixed bad reset vector in startup timing bug
Due to the async nature of fetching data, the 18 halfSteps performed on initChip were sometimes causing the system to read the reset address improperly.  For example, instead of going to $E000 it would go to $0000.  Still unsure exactly why (which is a bit bothering to not know), but initialization code is generally the most difficult to get right and to debug.

Also did some cleanup of functions that are not needed in this lighter weight version of the app.